# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: API RotaVan
  description: API para gestão de transporte escolar, incluindo cadastro de responsáveis, motoristas, escolas, rotas, veículos, contratos e viagens.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Servidor de Desenvolvimento Local
...
security:
  - BearerAuth: [] # Aplica segurança Bearer a todos os endpoints por padrão (exceto os explicitamente excluídos)
paths:
...
  /v1/api/veiculos:
    get:
      tags:
        - Veiculos
      summary: Lista todos os veículos.
      operationId: getAllVeiculos
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Veiculo'
        '401':
          description: Não autorizado.
    post:
      tags:
        - Veiculos
      summary: Cria um novo veículo.
      operationId: createVeiculo
      security:
        - BearerAuth: []
      requestBody:
        description: Dados do veículo a ser criado.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VeiculoInput'
      responses:
        '201':
          description: Veículo criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Veiculo'
        '400':
          description:
        '404':
          description: Motorista associado não encontrado.
        '401':
          description: Não autorizado.
  /v1/api/veiculos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID do veículo
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Veiculos
      summary: Busca um veículo pelo ID.
      operationId: getVeiculoById
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Veiculo'
        '404':
          description: Veículo não encontrado.
        '401':
          description: Não autorizado.
    put:
      tags:
        - Veiculos
      summary: Atualiza um veículo existente.
      operationId: updateVeiculo
      security:
        - BearerAuth: []
      requestBody:
        description: Dados do veículo atualizados.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VeiculoInput'
      responses:
        '200':
          description: Atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Veiculo'
        '400':
          description:
        '404':
          description: Veículo ou Motorista associado não encontrado.
        '401':
          description: Não autorizado.
    delete:
      tags:
        - Veiculos
      summary: Apaga um veículo.
      operationId: deleteVeiculo
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Apagado com sucesso.
        '404':
          description: Veículo não encontrado.
        '401':
          description: Não autorizado.
  /v1/api/veiculos/placa/{placa}:
    parameters:
      - name: placa
        in: path
        required: true
        description: Placa do veículo
        schema:
          type: string
    get:
      tags:
        - Veiculos
      summary: Busca um veículo pela placa.
      operationId: getVeiculoByPlaca
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Veiculo'
        '404':
          description: Veículo não encontrado.
        '401':
          description: Não autorizado.

  # ===============================================================
  # ENDPOINTS DE CONTRATOS (NOVOS)
  # ===============================================================
...
  /v1/api/viagens:
    get:
      tags:
        - Viagens
      summary: Lista todas as viagens.
      operationId: getAllViagens
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Viagem'
        '401':
          description: Não autorizado.
    post:
      tags:
        - Viagens
      summary: Cria uma nova viagem.
      operationId: createViagem
      security:
        - BearerAuth: []
      requestBody:
        description: Dados da viagem a ser criada.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ViagemInput'
      responses:
        '201':
          description: Viagem criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viagem'
        '400':
          description:
        '404':
          description: Criança, Motorista ou Rota não encontrada.
        '401':
          description: Não autorizado.
  /v1/api/viagens/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID da viagem
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Viagens
      summary: Busca uma viagem pelo ID.
      operationId: getViagemById
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viagem'
        '404':
          description: Viagem não encontrada.
        '401':
          description: Não autorizado.
    delete:
      tags:
        - Viagens
      summary: Apaga uma viagem.
      operationId: deleteViagem
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Apagada com sucesso.
        '404':
          description: Viagem não encontrada.
        '401':
          description: Não autorizado.
  /v1/api/viagens/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        description: ID da viagem
        schema:
          type: string
          format: uuid
    patch:
      tags:
        - Viagens
      summary: Atualiza apenas o status de uma viagem.
      operationId: updateViagemStatus
      security:
        - BearerAuth: []
      requestBody:
        description: Novo status da viagem.
        required: true
        content:
          application/json: # Pode ser text/plain se enviar só a string
            schema:
              type: string
              example: "Concluída"
      responses:
        '200':
          description: Status atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viagem'
        '400':
          description: Status inválido.
        '404':
          description: Viagem não encontrada.
        '401':
          description: Não autorizado.
  /v1/api/viagens/motorista/{motoristaId}:
    parameters:
      - name: motoristaId
        in: path
        required: true
        description: ID do Motorista
        schema:
          type: string
          format: uuid
      - name: data
        in: query
        required: true
        description: Data da viagem (formato YYYY-MM-DD)
        schema:
          type: string
          format: date
    get:
      tags:
        - Viagens
      summary: Busca viagens por motorista e data.
      operationId: getViagensByMotoristaAndData
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Viagem'
        '404':
          description: Motorista não encontrado.
        '401':
          description: Não autorizado.
  /v1/api/viagens/crianca/{criancaId}:
    parameters:
      - name: criancaId
        in: path
        required: true
        description: ID da Criança
        schema:
          type: string
          format: uuid
      - name: inicio
        in: query
        required: true
        description: Data de início do período (formato YYYY-MM-DD)
        schema:
          type: string
          format: date
      - name: fim
        in: query
        required: true
        description: Data de fim do período (formato YYYY-MM-DD)
        schema:
          type: string
          format: date
    get:
      tags:
        - Viagens
      summary: Busca viagens por criança e período.
      operationId: getViagensByCriancaAndPeriodo
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Viagem'
        '404':
          description: Criança não encontrada.
        '401':
          description: Não autorizado.


# ===============================================================
# COMPONENTS
# ===============================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Veiculo:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        placa: { type: string }
        capacidade: { type: integer }
        modelo: { type: string, nullable: true }
        motorista: { $ref: '#/components/schemas/MotoristaInfo' } # Simplificado
    VeiculoInput: # Para POST e PUT /veiculos
      type: object
      required: [ placa, capacidade, motoristaId ]
      properties:
        placa: { type: string }
        capacidade: { type: integer }
        modelo: { type: string, nullable: true }
        motoristaId: { type: string, format: uuid, description: ID do motorista existente }

    # Contrato
...
    Viagem:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        data: { type: string, format: date }
        horaSaida: { type: string, format: time, example: "06:30:00" } # Formato HH:mm:ss
        horaChegada: { type: string, format: time, nullable: true, example: "07:15:00" }
        status: { type: string, nullable: true }
        crianca: { $ref: '#/components/schemas/CriancaInfo' } # Simplificado
        motorista: { $ref: '#/components/schemas/MotoristaInfo' } # Simplificado
        rota: { $ref: '#/components/schemas/RotaInfo' } # Simplificado
    ViagemInput: # Para POST /viagens
      type: object
      required: [ data, horaSaida, criancaId, motoristaId, rotaId ]
      properties:
        data: { type: string, format: date }
        horaSaida: { type: string, format: time, example: "06:30:00" }
        horaChegada: { type: string, format: time, nullable: true, example: "07:15:00" }
        status: { type: string, nullable: true, example: "Agendada" }
        criancaId: { type: string, format: uuid, description: ID da criança existente }
        motoristaId: { type: string, format: uuid, description: ID do motorista existente }
        rotaId: { type: string, format: uuid, description: ID da rota existente }

    # --- Schemas Simplificados para Relacionamentos ---
    RotaInfo:
      type: object
      properties:
        id: { type: string, format: uuid }
        nomeRota: { type: string }
...
    MotoristaInfo:
      type: object
      properties:
        id: { type: string, format: uuid }
        nomeMotorista: { type: string }
...
    CriancaInfo:
      type: object
      properties:
        id: { type: string, format: uuid }
        nome: { type: string }